+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/web/PurchaseOrderProjectController.java
@@ -1,0 +1,43 @@
+/*
+ * Axelor Business Solutions
+ *
+ * Copyright (C) 2018 Axelor (<http://axelor.com>).
+ *
+ * This program is free software: you can redistribute it and/or  modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.axelor.apps.businessproject.web;
+
+import com.axelor.apps.purchase.db.PurchaseOrder;
+import com.axelor.apps.purchase.db.PurchaseOrderLine;
+import com.axelor.apps.purchase.db.repo.PurchaseOrderRepository;
+import com.axelor.exception.AxelorException;
+import com.axelor.rpc.ActionRequest;
+import com.axelor.rpc.ActionResponse;
+import com.google.inject.Inject;
+import com.google.inject.Singleton;
+
+@Singleton
+public class PurchaseOrderProjectController {
+
+  @Inject private PurchaseOrderRepository purchaseOrderRepo;
+
+  public void updateLines(ActionRequest request, ActionResponse response) throws AxelorException {
+    PurchaseOrder purchaseOrder = request.getContext().asType(PurchaseOrder.class);
+    purchaseOrder = purchaseOrderRepo.find(purchaseOrder.getId());
+
+    for (PurchaseOrderLine orderLine : purchaseOrder.getPurchaseOrderLineList()) {
+      orderLine.setProject(purchaseOrder.getProject());
+    }
+    response.setValue("purchaseOrderLineList", purchaseOrder.getPurchaseOrderLineList());
+  }
+}